pyproject.toml
[project]
name = "scalable-architecture-for-humans-and-ai"
version = "0.1.0"
description = "Provides scalable and discoverable architecture management for humans and ai"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "mcp>=1.27.0",
    "python-frontmatter>=1.1.0",
    "anyio>=4.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
    "ruff>=0.15.11",
    "zuban>=0.7.0",
]
gui = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
arch-gui-server = "src.infrastructure.gui.gui_server:main"
arch-backend = "src.infrastructure.backend.arch_backend:main"
arch-mcp-stdio-read = "src.infrastructure.mcp.arch_mcp_stdio:main"
arch-mcp-stdio-write = "src.infrastructure.mcp.arch_mcp_stdio_write:main_write"
arch-init = "src.infrastructure.workspace.workspace_init:main"
get-plantuml = "src.infrastructure.bootstrap.get_plantuml:main"
arch-model-write = "src.infrastructure.write.artifact_write_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = [
    ".arch",
    ".venv",
    "engagements",
    "enterprise-repository",
    "tests",
    "tools/gui",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]

[tool.zuban]
mode = "default"
python_version = "3.13"
files = [
    "src",
    "gen_id.py",
]
exclude = "(^|/)(\\.arch|\\.venv|engagements|enterprise-repository|tools/gui)(/|$)"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_unreachable = true

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
    "ruff>=0.15.11",
    "zuban>=0.7.0",
]
